requirements.txt
numpy
torch
matplotlib
opencv-python
tqdm
torchvision
PyYAML

